requirements.txt
fastapi==0.105.0
uvicorn==0.24.0.post1
python-dotenv==1.0.0
gradio-client==0.7.2
python-multipart==0.0.7